Add back old sweet.js API
