+++ .eslintignore
@@ -1,0 +1,1 @@
+test/test262-parser-tests
